don't use experimental-advance-branches
I can't find a way to be able to work directly in master or in branches with the same configuration.
When working in a branch for a PR with the bookmark set at the end of the branch, creating a new
commit to modify a commit already in the PR, and then squashing the new commit in the existing one,
the `master` bookmark is moved in the current branch.